backend/requirements-prod.txt
Django>=4.0
gunicorn==21.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
uwsgi
djangorestframework-simplejwt  # for login
djoser
python-dotenv
